requirements.txt
# PyTorch and related (install separately with CUDA support)
# pip3 install torch torchvision --index-url https://download.pytorch.org/whl/cu126

numpy>=1.21.0
scikit-learn>=1.0.0
matplotlib>=3.5.0

tqdm>=4.62.0
tensorboard>=2.8.0
torchsummary>=1.5.1